+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - the key function passed to `sortBy` may now also return a DateTime, a CalendarDate, or a bigint
 - fix `sort` on bigint arrays
 - `percent` function
+- `association` and `associationOrFail` methods to encapsulate the loaded check or associationQuery ternary pattern
 
 ## 1.9.4